+++ assets/jsx/workflow-editor/components/node-types/generic.jsx
@@ -230,8 +230,7 @@ export const GenericNode = memo(({ id, data, isConnectable, selected, nodeTypeIc
                 {targetHandles}
                 <div className='react-flow__node-top'>
                     {nodeTypeIcon}
-                    {topText}
-                    {/* {topText}: <span className='node-slug'>{data.slug}</span> */}
+                    {topText}: <span className='node-slug'>{data.slug}</span>
                 </div>
 
                 <div className={nodeDescription ? 'react-flow__node-inner-body with-description' : 'react-flow__node-inner-body'}>