+++ assets/jsx/workflow-editor/components/setting-popover/index.jsx
@@ -1,0 +1,36 @@
+import {
+    Popover,
+    __experimentalVStack as VStack,
+    __experimentalHStack as HStack,
+    Button,
+ } from "@wordpress/components";
+
+export const SettingPopover = ({ onClose, title, children, className, offset = 40, placement = "left-start", ...props }) => {
+    return (
+        <Popover
+            onClose={onClose}
+            placement={placement}
+            offset={offset}
+            className={className}
+            {...props}
+        >
+            <VStack>
+                <HStack>
+                    <h2 className="components-truncate components-text components-heading block-editor-inspector-popover-header__heading">
+                        {title}
+                    </h2>
+                    <Button
+                        icon={'no-alt'}
+                        isSmall={true}
+                        className="block-editor-inspector-popover-header__action"
+                        onClick={onClose}
+                    />
+                </HStack>
+            </VStack>
+
+            {children}
+        </Popover>
+    );
+};
+
+export default SettingPopover;