+++ CHANGELOG.md
@@ -9,12 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add new trigger: Post Author Changed - PRO (Issue #1144).
 - Add new action: Update Post - PRO (Issue #1143).
+- Add the step slug/name to the top of the step's node in the workflow editor.
 
 ### Changed
 
 - Changed the description field on workflow steps settings to be a popover, and part of the inspector card, saving space in the sidebar.
 - Remove left padding from the workflow step details panel.
-- Remove attributes table from the inspector card when on developer mode (that data can still be seen in the Developer Info panel).
+- Remove attributes table from the inspector card when on developer mode (that data can still be seen in the Developer Info panel)
 
 
 ## [4.3.3] - UNRELEASED